Fixed the issue: 19601

+++ dotnet/corecomponents/net.atos.daf.ct2.reports/repository/ReportRepository_FleetFuel.cs
@@ -550,11 +550,11 @@ namespace net.atos.daf.ct2.reports.repository
 				  , 1 as totalworkingdays
 				  , (etl_gps_distance)                                                  as etl_gps_distance
 				  , (etl_gps_distance)                                              as veh_message_distance
-				  , (etl_gps_distance)/case when (etl_gps_trip_time) >0 then (etl_gps_trip_time) else 1 end as average_speed
+				  , (etl_gps_distance::numeric)/case when (etl_gps_trip_time::numeric) >0 then (etl_gps_trip_time) else 1 end as average_speed
 				  , (max_speed)                                                         as max_speed
 				  , (average_gross_weight_comb)                                      as average_gross_weight_comb
 				  , (etl_gps_fuel_consumed)                                             as fuel_consumed
-				  , ((etl_gps_fuel_consumed)/case when (etl_gps_distance) >0 then (etl_gps_distance) else 1 end) as fuel_consumption
+				  , ((etl_gps_fuel_consumed::numeric)/case when (etl_gps_distance::numeric) >0 then (etl_gps_distance) else 1 end) as fuel_consumption
 				  , (co2_emission)                                                      as co2_emission
                   , idle_duration as idle_duration
 				  , case when (((end_time_stamp - start_time_stamp)/1000)::numeric)>0 then ((idle_duration/((end_time_stamp - start_time_stamp)/1000)::numeric) *100) else 0 end as idle_duration_percentage
